+++ Controls-demo/explorerNew/BreadCrumbsInHeader/Default/Index.ts
@@ -5,6 +5,7 @@ import { IColumn } from 'Controls/grid';
 import {TRoot} from 'Controls-demo/types';
 import { IHeaderCell } from 'Controls/grid';
 import {HierarchicalMemory} from 'Types/source';
+import {IItemAction} from 'Controls/itemActions';
 
 export default class extends Control {
     protected _template: TemplateFunction = Template;
@@ -13,6 +14,7 @@ export default class extends Control {
     protected _viewMode: string = 'table';
     protected _root: TRoot = 1;
     protected _header: IHeaderCell[] = Gadgets.getHeader();
+    protected _itemActions: IItemAction[];
 
     protected _beforeMount(): void {
         this._viewSource = new HierarchicalMemory({
@@ -20,6 +22,21 @@ export default class extends Control {
             parentProperty: 'parent',
             data: Gadgets.getData()
         });
+
+        this._itemActions = [
+            {
+                id: 1,
+                icon: 'icon-PhoneNull',
+                title: 'phone',
+                showType: 2
+            },
+            {
+                id: 2,
+                icon: 'icon-EmptyMessage',
+                title: 'message',
+                showType: 2
+            }
+        ];
     }
 
     static _styles: string[] = ['Controls-demo/Controls-demo'];
