--- Controls-demo/explorerNew/ItemActions/Base/Base.wml
@@ -1,24 +1,0 @@
-<div class="controlsDemo__wrapper">
-    <Controls.browser:Browser source="{{_viewSource}}" keyProperty="id" bind:filter="_filter"
-                              searchParam="title" minSearchLength="{{3}}" searchStartingWith="{{_searchStartingWith}}">
-        <div class="ws-fragment">
-            <Controls.search:InputContainer>
-                    <Controls.search:Input/>
-            </Controls.search:InputContainer>
-            <Controls.list:Container>
-                <Controls.explorer:View
-                    source="{{_viewSource}}"
-                    itemActions="{{_itemActions}}"
-                    bind:root="_root"
-                    attr:class="demo-Explorer ControlsDemo-Explorer"
-                    keyProperty="id"
-                    displayProperty="title"
-                    parentProperty="parent"
-                    nodeProperty="parent@"
-                    columns="{{_columns}}">
-                </Controls.explorer:View>
-            </Controls.list:Container>
-        </div>
-    </Controls.browser:Browser>
-</div>
-
