--- Controls-demo/explorerNew/ItemActions/OutsideWithBreadCrumbs/OutsideWithBreadCrumbs.wml
@@ -1,17 +1,0 @@
-<div class="controlsDemo__wrapper">
-    <Controls.explorer:View
-            source="{{_viewSource}}"
-            bind:root="_root"
-            attr:class="demo-Explorer ControlsDemo-Explorer"
-            itemActions="{{_itemActions}}"
-            itemActionsPosition="outside"
-            keyProperty="id"
-            displayProperty="title"
-            parentProperty="parent"
-            nodeProperty="parent@"
-            viewMode="{{_viewMode}}"
-            rowSeparatorSize="s"
-            header="{{_header}}"
-            columns="{{_columns}}">
-    </Controls.explorer:View>
-</div>
